+++ modules/core/23-commitment/types/merkle.go
@@ -9,7 +9,7 @@ import (
 
 	errorsmod "cosmossdk.io/errors"
 
-	tmcrypto "github.com/cometbft/cometbft/proto/tendermint/crypto"
+	cmtcrypto "github.com/cometbft/cometbft/proto/tendermint/crypto"
 
 	"github.com/cosmos/ibc-go/v8/modules/core/exported"
 )
@@ -243,7 +243,7 @@ func verifyChainedMembershipProof(root []byte, specs []*ics23.ProofSpec, proofs 
 // and are declared as globals to avoid having to unnecessarily re-allocate on every comparison.
 var (
 	blankMerkleProof = &MerkleProof{}
-	blankProofOps    = &tmcrypto.ProofOps{}
+	blankProofOps    = &cmtcrypto.ProofOps{}
 )
 
 // Empty returns true if the root is empty